requirements.txt
click>=7.0
pytorch-ignite~=0.4.9
tensorboard>=2.5
scipy==1.3.1
numpy==1.19.5
pillow==6.2.1
timm~=0.4.12
